backend/requirements.txt
aiocache==0.11.1
aiofiles==22.1.0
aiohttp==3.8.3
aiopg==1.4.0
aiosignal==1.2.0
alembic==1.8.1
anyio==3.6.2
asgi-lifespan==2.0.0
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
camel-converter==2.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.3
coverage==6.5.0
cryptography==38.0.3
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.4
fastapi==0.85.0
fastapi-cache2==0.1.9
fastapi-discord==0.2.4
Flask==2.0.2
frozenlist==1.3.1
greenlet==2.0.0
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
imgbbpy==0.1.3
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.2.3
MarkupSafe==2.1.1
meilisearch==0.22.2
meilisearch-python-async==0.28.1
multidict==6.0.2
numpy==1.23.4
outcome==1.2.0
packaging==21.3
peewee==3.15.3
peewee-async==0.8.0
pendulum==2.1.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.2
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
pytzdata==2020.1
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlmodel==0.0.8
starlette==0.20.4
tomli==2.0.1
trio==0.22.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
yarl==1.8.1
